chore: Refactor helper function of management command `createsuperuser`
Ref: https://cordada.aha.io/features/COMPCLDATA-112
